Add icon image and fix screen sizing
Added an icon image to be displayed on the application.

Added new resources directory.

Moved launch4j icon into resources directory.

Fixed screen sizing so it adjusts correctly based on the users
resolution.

Fixed interface location so that it appears in the centre of the screen
rather than the corner on startup.